@@ -4,25 +4,29 @@
   "private": true,
   "dependencies": {
     "ramda": "^0.25.0",
-    "ramda-adjunct": "^2.6.0",
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
+    "ramda-adjunct": "^2.10.0",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
     "react-redux": "^5.0.7",
-    "react-router": "^4.2.0",
-    "react-scripts": "1.1.4",
+    "react-router": "^4.3.1",
+    "react-scripts": "^2.0.5",
     "redux": "^3.7.2",
     "redux-devtools-extension": "^2.13.2",
-    "redux-observable": "^0.18.0",
-    "rxjs": "^5.5.10",
-    "styled-components": "^3.2.5"
+    "redux-observable": "^1.0.0",
+    "rxjs": "^6.3.3",
+    "styled-components": "^4.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "format": "prettier-standard 'src/**/*js'",
-    "precommit": "lint-staged",
     "eject": "react-scripts eject"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
   },
   "lint-staged": {
     "linters": {
@@ -33,16 +37,16 @@
     }
   },
   "devDependencies": {
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "enzyme-to-json": "^3.3.3",
-    "husky": "^0.14.3",
-    "jest-enzyme": "^6.0.0",
-    "jest-styled-components": "^5.0.1",
-    "lint-staged": "^7.0.4",
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.6.0",
+    "enzyme-to-json": "^3.3.4",
+    "husky": "^1.1.2",
+    "jest-enzyme": "^7.0.0",
+    "jest-styled-components": "^6.2.1",
+    "lint-staged": "^7.3.0",
     "prettier-standard": "^8.0.1",
-    "react-test-renderer": "^16.3.1",
-    "redux-mock-store": "^1.5.1"
+    "react-test-renderer": "^16.5.2",
+    "redux-mock-store": "^1.5.3"
   },
   "jest": {
     "collectCoverageFrom": [
@@ -58,5 +62,11 @@
     "snapshotSerializers": [
       "enzyme-to-json/serializer"
     ]
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
